requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.10
aiosignal==1.4.0
alabaster==1.0.0
altair==5.5.0
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
babel==2.17.0
backports.functools-lru-cache==2.0.0
beautifulsoup4==4.14.3
bitsandbytes==0.48.2
blinker==1.9.0
cachetools==6.2.1
cbor==1.0.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
datasets==4.3.0
debugpy==1.8.17
decorator==5.2.1
dill==0.4.0
distro==1.9.0
docutils==0.21.2
emoji==2.15.0
environs==9.5.0
executing==2.2.1
fastapi==0.118.0
filelock==3.20.0
FlagEmbedding==1.3.5
Flask==2.1.3
flask-cors==6.0.1
flask-restx==1.3.2
flatbuffers==25.9.23
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.9.0
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
ijson==3.4.0.post0
imagesize==1.4.1
importlib_resources==6.5.2
indic-nlp-library==0.92
indic-transliteration==2.3.75
iniconfig==2.3.0
inscriptis==2.7.0
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
ir_datasets==0.5.11
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
langchain-core==1.0.3
langchain-text-splitters==1.0.0
langdetect==1.0.9
langsmith==0.4.40
lxml==5.4.0
lz4==4.4.5
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.1.0
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
Morfessor==2.0.6
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
narwhals==2.11.0
neo4j==6.1.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.5
onnx==1.20.0
onnxruntime==1.23.2
onnxruntime-tools==1.7.0
openai==1.55.3
optimum==2.0.0
optimum-onnx==0.0.3
orjson==3.11.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
peft==0.18.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
prometheus_client==0.20.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py3nvml==0.2.7
pyarrow==21.0.0
pydantic==2.11.9
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.1
Pygments==2.19.2
pymilvus==2.6.5
PyMuPDF==1.26.4
pyparsing==3.2.5
pyreadline3==3.5.4
pytesseract==0.3.13
pytest==8.3.5
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
roman==5.2
roman-numerals-py==3.1.0
rpds-py==0.28.0
safetensors==0.6.2
sanskrit-parser==0.2.6
sanskrit-util==0.1.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
sentence-transformers==5.1.2
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.8.3
Sphinx==8.2.3
sphinx-argparse==0.5.2
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.44
stack-data==0.6.3
stanza==1.11.0
starlette==0.48.0
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tinydb==4.8.2
tokenizers==0.21.0
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.55.0
trec-car-tools==2.6
typer==0.15.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
Unidecode==1.4.0
unlzw3==0.2.3
urllib3==2.5.0
uuid_utils==0.14.0
uvicorn==0.37.0
vidyut==0.4.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchdog==6.0.0
wcwidth==0.2.14
Werkzeug==2.1.2
wget==3.2
xlrd==2.0.2
xmltodict==1.0.2
xxhash==3.6.0
yarl==1.22.0
zlib-state==0.1.10
zstandard==0.25.0